+++ src/DesktopVirtualization/DesktopVirtualization.Autorest/test/Update-AzWvdHostPool.Tests.ps1
@@ -40,7 +40,7 @@ Describe 'Update-AzWvdHostPool' {
                 $hostPool.FriendlyName | Should -Be 'fri'
                 $hostPool.MaxSessionLimit | Should -Be 5
                 $hostPool.VMTemplate | Should -Be '{option1}'
-                # @todo not corrct since it should be null need to look into it
+                # @todo not correct since it should be null need to look into it
                 # $hostPool.CustomRdpProperty | Should -Be ""
                 $hostPool.Ring | Should -Be $null
                 # @todo need to check this
@@ -69,7 +69,7 @@ Describe 'Update-AzWvdHostPool' {
                 $hostPool.FriendlyName | Should -Be 'fri2'
                 $hostPool.MaxSessionLimit | Should -Be 6
                 $hostPool.VMTemplate | Should -Be '{option2}'
-                # @todo not corrct since it should be null need to look into it
+                # @todo not correct since it should be null need to look into it
                 # $hostPool.CustomRdpProperty | Should -Be ""
                 $hostPool.Ring | Should -Be $null
                 # @todo need to check this
@@ -88,7 +88,7 @@ Describe 'Update-AzWvdHostPool' {
                 $hostPool.FriendlyName | Should -Be 'fri2'
                 $hostPool.MaxSessionLimit | Should -Be 6
                 $hostPool.VMTemplate | Should -Be '{option2}'
-                # @todo not corrct since it should be null need to look into it
+                # @todo not correct since it should be null need to look into it
                 # $hostPool.CustomRdpProperty | Should -Be ""
                 $hostPool.Ring | Should -Be $null
                 # @todo need to check this