+++ src/Storage/Storage/Blob/Cmdlet/GetAzureStorageBlobContent.cs
@@ -54,7 +54,7 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Blob.Cmdlet
         private const string ContainerParameterSet = "ContainerPipeline";
 
         /// <summary>
-        /// downlaod from uri parameter set
+        /// download from uri parameter set
         /// </summary>
         private const string UriParameterSet = "UriPipeline";
 
@@ -445,7 +445,7 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Blob.Cmdlet
                 throw new ArgumentException(String.Format(Resources.InvalidFileName, fileName));
             }
 
-            //there is no need to check the read/write permission on the specified file path, the datamovement libraray will do that
+            //there is no need to check the read/write permission on the specified file path, the data movement library will do that
 
             return filePath;
         }