+++ src/LambdaTest/LambdaTest/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* General availability for module Az.LambdaTest
+* Upgraded the API version to 2024-02-01
 
 ## Version 0.1.0
 * First preview release for module Az.LambdaTest